feat(examples): add simpler slash-commands example
